fix potential blocking of NettyTransport connect and disconnect methods
The previous locking mechanism (synchronize(connectLock(String))) required
that id of the DiscoveryNode for two nodes did not map to the same entry in
the array connectMutex. If this was not the case, this was likely to cause a
deadlock between methods acquiring locks using the master id and the
node id.
This very rarely caused timeouts in
SpecificMasterNodesTests#simpleOnlyMasterNodeElection.

Instead it shoud be assured that different ids are mapped to different
locks.