Make sure that selected terms that are not on doc are returned with freq 0
If specific terms are requested, they should be returned with freq 0 if they do not
appear in the document. To do this efficiently, we compute a sorted list of
terms and then intersect this list with the terms in the documents when writing
the terms.

This commit changes an inconsistency of uri parameter handling:
Previously, when selected fields were given in the uri parameters,
these fields were added to the list of selected terms. This does not make
sense since it is not the other way round. Now, uri parameters are always
overwritten by the parameters given in the body.

closes #3924 for single term vector request