Add parameter to GET for checking if generated fields can be retrieved
Fields of type `token_count` are generated only when indexing.
If a GET requests accesses the transaction log (because no refresh
between indexing and GET request) and the token_count field is requested,
then GET will try to parse the number from the String and the result
is a NumberFormatException.
Now, GET accepts a parameter `ignore_errors_on_generated_fields` which has
the following effect:
- Throw exception with meaningful error message explaining the problem if set to false (default)
- Ignore the field if set to true

closes #6676