[store] Allow shard deletion also if shards on other nodes are in POST_RECOVERY
When we check on other nodes if the expected number of copies is available before
deleting the directory, but the one copy is still in POST_RECOVERY the we do not delete
the shard. However, we might run into this situation in case the shard has just been
relocated from one node to another.
The test IndicesStoreIntegrationTests.indexCleanup fails occationally because of this.